Added server actions security and modal page scroll update

+++ actions/items/deleteItem.js
@@ -1,16 +1,46 @@
 'use server';
 
 import Task from '../../models/Task';
+import { cookies } from 'next/headers';
+import { jwtVerify } from 'jose';
 import { handleServerErrorMessage } from '../../utilities';
+const jwtSecret = process.env.JWT_SECRET;
 
-export default async function deleteTask(_id) {
+export default async function deleteTask(_id, userId) {
+  // check that cookie user id matches param userId
+  const token = (await cookies()).get('saturday');
+  let cookieUserId;
+
+  if (token) {
+    try {
+      const { payload } = await jwtVerify(
+        token?.value,
+        new TextEncoder().encode(jwtSecret)
+      );
+      if (payload?.id) {
+        cookieUserId = payload?.id;
+      }
+    } catch (error) {
+      const errorMessage = handleServerErrorMessage(error);
+      console.error(errorMessage);
+      return { status: 500, error: errorMessage };
+    }
+  }
+
+  if (!userId || userId !== cookieUserId) {
+    return {
+      status: 400,
+      error: 'Unauthorized',
+    };
+  }
+
   try {
     const result = await Task.find({ _id: _id });
     await Task.deleteOne({ _id: _id });
     return { status: 200, item: JSON.parse(JSON.stringify(result[0])) };
   } catch (error) {
-    console.log(error);
     const errorMessage = handleServerErrorMessage(error);
+    console.error(errorMessage);
     return { status: 500, error: errorMessage };
   }
 }
